fixup! refactor(login-strategy): centralize setAccountCryptographicState implementation
